+++ .bumpversion.cfg
@@ -4,7 +4,7 @@ tag_name = {new_version}
 commit = False
 tag = False
 
-[bumpversion:file:alfred/__init__.py]
+[bumpversion:file:robin/__init__.py]
 search = version="{current_version}"
 replace = version="{new_version}"
