+++ .isort.cfg
@@ -3,4 +3,4 @@ line_length = 88
 force_single_line = True
 lines_after_imports = 2
 force_sort_within_sections = True
-known_first_party = alfred
+known_first_party = robin
