+++ pytest.ini
@@ -4,4 +4,4 @@ testpaths = tests
 norecursedirs = _*
 timeout=30
 timeout_method=thread
-cache_dir = /tmp/.alfred.pytest.cache
+cache_dir = /tmp/.robin.pytest.cache
